refactor(sliceview/chunked_graph): Remove VolumeChunk/ChunkFormatHandler dependency
